pyproject.toml
[tool.poetry]
name = "nonebot-plugin-githubcard"
version = "0.4.0"
description = "检测GitHub仓库链接并自动发送卡片信息，支持仓库、Issue、Pull Request、Release"
authors = ["Nanako <demo0929@vip.qq.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ElainaFanBoy/nonebot_plugin_githubcard"
repository = "https://github.com/ElainaFanBoy/nonebot_plugin_githubcard"
documentation = "https://github.com/ElainaFanBoy/nonebot_plugin_githubcard#readme"
keywords = ["nonebot", "github", "card", "bot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Robot Framework :: Library",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Communications :: Chat",
]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0"
nonebot-adapter-onebot = "^2.2.0"
nonebot-plugin-alconna = "^0.30.0"
aiohttp = "^3.7.4"
pydantic = "^1.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
black = "^22.0.0"
isort = "^5.10.0"
flake8 = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
